Implement exploration for ngff v0.5
